Football_news/HTML_News/news_html/requirements.txt
fastapi[standard]
uvicorn[standard]
requests
bs4
lxml_html_clean
newspaper3k
